🔧 Establish a "build" step in our workflow
In this commit, we add a "build" script to the package.json, which
we can invoke via `yarn build`. Currently, this script simply copies
the contents of the **src/** directory to a **lib/** directory.

By doing this now, we establish a "build" step into our workflow that
indicates that our code may undergo some kind of transformation prior
to being published. This means that:

 1. No files outside of the **src/** directory should reference
    any files within it. Instead, those files should target the
    **lib/** directory.

 2. Files within the **src/** directory **may** technically work
    in an environment without any transformations, but because
    these files are not to be referenced directly, we are free to
    develop free of the constraints of any particular environment,
    so long as we have the tools to transform that code into a
    portable form when it is published.

For now, I have introduced a small "Hello world" stub that should run
in any environment. Since no transformations are required, the build
step is a simple copy operation.