🔧 Migrate "build" step to Webpack
Webpack is a versatile build tool for JavaScript projects.
It can produce a portable, optimized file from source code
using non-portable syntax and features. The most common
example of modern syntax that is not consistently supported
across platforms is ES6 imports. This feature is **essential**
for any modern JavaScript developer, and unfortunately does
not have consistent cross-platform support (yet).

There are other "bundlers" and "transpilers" available, like
Parcel and Babel (respectively). If later we learn that Webpack
is inadequate for this project's portability, and module resolution,
and/or optimization needs, we can try another tool or set of tools.