🔧 Configure Webpack for "production" mode
We want to get into the habit early of having development match
production as closely as possible. This can help to prevent toolchain
issues locally, before even triggering a CI build and well before
deploying that code to production.

In this commit, we're just passing in `--mode production` to Webpack
in the "build" script, which opts in to production-optimized defaults.