🔧 Add ESLint for static analysis and to enforce code style
ESLint is a tool that helps to automatically enforce code quality
and style. By adding it into the toolchain early in the project's
development, common errors can be detected (and prevented!), which
accelerates development time by minimizing painful debugging sessions.

For ESLint to be most effective, it needs to know a little bit about
the environment(s) in which the app is designed to run. This is one
of the decision points that will vary for your project based on whether
you are building a web app (client, front-end), a web API (server,
back-end), or a reusable package.

I'm electing in this commit to make as few decisions as is necessary
in order to successfully introduce a *lint* step to the workflow.

 1. First, I added the `eslint@6.6.0` dev dependency.

 2. Next, I ran `npx eslint --init`, which prompts with a questionairre
    where I asked some general questions about this project. Based on
    my answers, that prompted for installation of a few more dependencies.

 3. Because we're using Yarn in this project (vs NPM), and the ESLint
    questionairre assumes NPM, I had to clean up the **package-lock.json**.
    Because we're using specific version selectors (vs fuzzy selectors),
    I had to remove the preceding `^` from the versions of the dependencies
    added by ESLint. Then, I had to run `yarn install` again to update
    the Yarn lockfile.

 4. Finally, I reformatted the **.eslint.json** configuration file to
    indent with two spaces instead of four spaces.

Now would be a good opportunity to talk about **version selectors**.

The majority of reusable packages in the npm registry use "fuzzy"
version selectors. This means that when a package manager like NPM or
Yarn resolves that package's dependencies, it will choose the most
recent version of that package, as long as it matches the **major**
version of the fuzzy version selector. This technique relies heavily
on dependency packages consistently following **semantic versioning**
and consistently avoiding **regressions** in existing functionality.

With so many packages depending on others, it doesn't take much for
the dependency graph to explode to hundreds, thousands, or even tens
of thousands of packages. Before *lock files*, this resulted in
extensive variability in dependency resolutions between each attempt,
even minutes apart. To solve for this, Yarn introduced **yarn.lock**,
and eventually NPM introduced **package-lock.json**, to "lock" these
dependency resolutions such that they resolve consistently from build
to build.

Lock files help to solve the problem, but it requires committing a
tool-generated file that can be extremely difficult for a human to
interpret, and impossible for a human to write. These files can be
enormous. So large that many peer review tools simply **ignore** them
or suppress the diff. Among the trade-offs of using lock files is
that human reviewers have become accustomed to ignoring changes to
them, yet they can be manipulated into false resolutions that a human
reviewer would never catch due to the overwhelming abundance of noise
that lock files generate.

Lock files are an artifact of the abundance of packages in the
JavaScript ecosystem that use fuzzy version selectors instead of
specifying the exact versions of the packages they depend on.

By declaring specific versions of dependencies in **package.json**,
the package's author establishes a guarantee that the package was
designed to work with that version of that dependency. Wherever
that package is used, its module resolutions will match.