🎨 Commit to this being a browser app
Previously, the `console.log()` statement was triggering the `no-console`
ESLint rule. This commit takes the first step toward introducing DOM
interactions into the source code.

The `#mount()` function introduced in this commit is the first *module*
of the package. Currently, it just injects a `<div>` element into the
document's body, removing the previous one if necessary, and appends
the given node to that element.

Ah, but don't take my word for what `#mount()` does.
I'll prove it with some tests.