🎨 Add unit tests for #mount() helper function
There are many choices of frameworks to help writing tests be
more joyful, or at least less painful. A test suite should be
fluid, fun to write, easy to read, easy to run, fast, and most
importantly, concise! Too much abstraction and code reuse and
individual tests become less clear. Too little abstraction and
tests start to contain too much boilerplate. Writing good tests
involves striking a balance between the two extremes, and
remaining focused on writing *each* test with brevity and
intention.

Mocha is a testing framework that follows a pattern like:

    describe('<name of the unit>', () => {
      it('does this specific thing', () => {
        // Exercise the code under test and verify that specific thing
      });
    });

Chai is an assertion library with a few different variants,
but I have selected the `expect` variant, which follows a pattern
like:

    expect(car).to.have.property('make', 'Ford');

To run the test suite against the pre-built source code, the Mocha
test runner needs to be able to parse the modern syntax we're using
in the source code. As an added benefit, by enabling this, we can
use a consistent syntax in our test suite as well.

Babel enables modern JavaScript syntax, much like Webpack does for
the source code. In fact, Webpack is already using Babel to do this.
The only difference now is that we're using Babel directly, too, for
processing the test suite.

Finally, because the `#mount()` function we want to test involves
DOM manipulation, but we want to run the test suite within a
browserless environment, we're using JSDOM as a lightweight
alternative implementation of DOM standards. While this does not
accurately reflect the many nuances and eccentricities of all
browser implementations of the DOM, it can provide a close enough
approximation for all DOM manipulations we care to test for.

I've also had to modify the `#mount()` function to accommodate
being run in an environment without browser-only globals (in this
case, `document`). This is one example of how **testability impacts
software design**. Untested code may have implicit global dependencies
that can be tricky or impossible to mock. If a unit is mutating
one of these implicit globals, this can be referred to as "side
effects". It is very common (and necessary) for a unit to mutate its
arguments, such as in `#mount()` whose goal is to mount the given
element into the given document. Adding the `document` to which the
element should be mounted is a natural expression of what the unit
does.

This pattern of refactoring implicit global dependencies into
arguments passed into a unit is called **dependency injection**.