🐛 Fix `#mount()` and its test suite
In the previous commit where we introduced a test suite for the
`#mount()` function, the test suite did not do what we thought it
did. All the tests were passing, but the expectations were incorrect.

To the observant reviewer, this may have been apparent.

See, `#mount()` was not injecting an intermediate `<div>` to the
document body. It was removing a pre-existing `<div>`, but would
inject the given element directly into the body. So, why did the
test suite pass?

 1. The test for whether a `<div>` was injected into the body was
    testing only for `undefined`, but `#querySelector()` was
    returning `null`. This was not undefined, so it satisfied the
    expectation as written.

 2. The other tests were using the `.include()` assertion, which
    always seems to pass when not acting upon an array.

So all three of our tests were giving us false positives.

When writing a unit test **after** writing the unit itself, it is
very common to end up with false positives due to errors in how the
test itself is written. This is why many seasoned developers
recommend a "test first" approach, or TDD.

With TDD, development occurs with a "red/green/refactor" cycle.

 * **Red**: Write a unit test that expresses some expectation you
   want the unit to satisfy. It should fail because you have not
   written the code, yet.

 * **Green**: Write the code in the unit under test to satisfy the
   expectation you just wrote. Write the minimum amount of code to
   satisfy that test, and stop as soon as the test passes (green).

 * **Refactor**: Continue to refactor the code until you are
   satisfied with the implementation, making sure all tests
   continue to pass.

This cycle should typically take a few minutes or so, end to end.

Following this technique ensures that your test suite contains no
false positives, because the test will always have been, at one
point, in a "red" state.

-----

Additionally, we had some lint errors in the previous commit.
All reported errors were related to how test code, specifically, is
written:

 * The test suite imports from devDependencies, by design.

 * Expectations in the test suite may appear as expressions, which
   ESLint reports as "unused expressions".

Both of these rules are relevant to the source code, but do not
apply to the test suite, where these two patterns are common and
intentional.

Thus, the ESLint configuration (.eslintrc.json) has been modified
with overrides for these two rules that apply only to tests.

-----

One more general note about paths in this project.
**index.js** is generally resolved as the "default" path when
the module resolver is given a path to a directory. We exploit this
to allow us to place the test suite for each module directly alongside
the implementation. This has some other benefits, as well:

 * Other support files, like a README, stylesheet, mocks, etc, can
   be placed directly alongside the implementation of the module
   itself.

 * Each module can be extracted as a separate top-level package,
   if desired, simply by copying that module's directory.

Typically, a module directory uses *kebab-case*, the implementation
is in **index.js**, and the test suite is in **spec.js**.
We use this pattern, like in the ESLint configuration, to easily
target subsets of the files in the repo based on their role.