🔧 Start tracking code coverage
In the previous commit, I talked about some of the benefits of test
driven development (TDD). One benefit I did not mention is **code
coverage**.

Code coverage is a measurement of how much of the source code is
being exercised by the test suite. In other words, it can tell you
if you have any **untested code**. If the team is practicing TDD,
then **untested code is effectively dead code**.

Well, we can *measure* code coverage. Istanbul (`nyc`) is one such
tool for measuring this. If we can measure it, then we can **enforce**
it, too.

Measuring (and enforcing) code coverage is great, but I want to clarify
something: **100% code coverage does not imply 100% test coverage**.
Basically, just because your test suite exercises every line of code,
this does not prove that your test suite covers all *boundary cases*
and *classes of inputs*. This is still the responsibility of the human
developer to write, and of the human reviewer to enforce.