📝 Add a license
In the initial commit, I had already specified that this project would
use the MIT license. I've found this to be a generally good license to
use for open-source projects, and a reasonable default to use if you
are just getting started with licensing.

Even with a license declaration in **package.json**, though, it is
generally a best practice to include the test of the license you have
chosen in the repo itself, as **LICENSE**, **LICENSE.txt**, or similar.

This commit introduces that.