🔧 Add step to generate code-level documentation
Documentation can be a divisive subject amongst developers.

 * Some developers consider code-level documentation to be
   a waste of time.

 * Some developers consider code-level documentation to be
   an essential part of developer and project hygiene.

The truth is, both opinions have merits. As with any practice
you adopt as a developer, you should always be asking yourself,
"Why am I doing this?" and "What is the value of doing this?"

Investing your time on creating and maintaining good code-level
documentation is only worth it if you, or other developers, are
**reading** that documentation. A close-knit team working in a
single project together may know the code well enough that the
documentation may seem pointless. Also, if there is a chance that
documentation may drift and become inaccurate or untrustworthy,
then even new developers onboarding onto the project may shy away
from it as a precaution against being potentially misinformed.

If you are developing a reusable package, then you almost
certainly should be writing code-level documentation on, at least,
the parts of your code you expect other developers to use. This
is how most developers will decide whether to use your package,
and will service as a perpetual reference on *how* they exercise it.

If you are developing an app (client, server, or otherwise), then
the benefits of code-level documentation are less clear. Since
other developers are not consuming your code, the audience for your
documentation is other developers *contributing* directly to the
app. These developers have access to the source code itself, and
are capable of navigating its implementation and/or consulting
directly with the other app developers to share knowledge. In this
case, the practice of writing and maintaining code-level documentation
is a decision best left to the team, based on its merits to how
the team likes to work.

When I refer to documentation in this context, by the way, I mean
API reference documentation. Typically, for each documented function:

 * A brief description of what the function is designed to accomplish.

 * A list of the arguments expected by the function, including each
   argument's name, type, and a brief explanation.

 * A description of what the function returns, including the return
   type.

JSDoc is a common way of describing functions and other units with
code-level block comments, so this commit integrates that.

To illustrate what this documentation looks like, I've documented
the `#mount()` function. I've also changed the name of its `element`
argument to `node`, since this argument can be any DOM node, not just
an element.

Finally, I've added a .gitignore entry for **/docs/**, which is where
we've elected to write the generated documentation from JSDoc.