💓 Add my Github Sponsorship details
If you like this project, please consider sponsoring me on Github!

> https://github.com/sponsors/canterberry

Take a look at some of my other projects, too:

> https://github.com/canterberry

I've been a proud contributor to open-source software on Github
since April 2013. I love what I do, and want to give back as best
as I can to the community not only by building and contributing to
quality OSS projects, but also by sharing the knowledge and
experience I have accumulated over the years as a software
professional with the next generation of developers. Together,
we can -- and will -- build great things.