📝 Describe, in the README, what this project really is
This commit summary, along with the changeset, says it all.

Let me take this opportunity, then, to talk about how I structure
these commit messages. You may have noticed a pattern. I'll admit,
I did not come up with this pattern on my own.

> See: https://github.com/atom/atom/blob/master/CONTRIBUTING.md#git-commit-messages

I try to keep my commit messages in conformance, roughly, with the
Atom (Editor) contribution guidelines. This is not due to any
particular affinity to that project. It just resonates with me as
being a really nice way of writing commit messages:

 * The emoji in the first line indicates what kind of change the
   commit represents. This can sometimes be ambiguous because the
   nature of each commit does not always fall clearly into one
   category or another.

 * I try to keep each line to 72 characters or less. This tends to be
   about where most truncation occurs in common UI components where
   commit messages tend to be viewed.

 * I use present-tense, second-person perspective. After all, I
   (the author) am writing directly to you (the reader).

Personally, I find commit messages to be a very rich source of
information about both a project and its contributors. So much
context *can* live in commits. It's the one place where a developer
can describe an entire atomic changeset and their thoughts and
intentions behind that commit. The discipline to make frequent
atomic commits with a cohesive purpose is a strong positive signal
of building with *mindfulness* and *intent*. There, I said it again!