📦 Upgrade dependencies to their latest versions
It's generally a good idea to check your dependencies from time to time
to see if there are any updates available. `yarn outdated` is an easy
way to do it.

There are some services, like Greenkeeper and Snyk, that can monitor
your dependencies for you. Some will even automatically create a pull
request for you to review, triggering your CI build which runs your
test suite to ensure upgrading that dependency upgrade is compatible
with your app or library.

Alright, but **why**? What is the value of keeping dependencies up
to date? If you're not actively and immediately planning to use a
new feature only available in the new version, then why bother?

 1. Upgrading a dependency may introduce defects, subtle compatibility
    issues, and other regressions that were not previously present.

 2. The rate at which dependency upgrades are available increases,
    at least, linearly with the number of direct dependencies your
    project has. This number is much higher if you take transitive
    dependencies into consideration. A higher rate of potential
    upgrades means higher maintenance overhead and higher risk.

Alright, so those are some strong points against doing this. Here
are some arguments **for** regular dependency upgrades:

 1. Eventually, there may be a critical update to one or more of
    your dependencies. By upgrading often, there is a smaller gap
    between the version your app is using and the version in which
    the defect is fixed.

 2. Documentation and support for most open-source packages is
    limited to the most recent version only. When you need to file
    an issue, often the first thing you will need to do is provide
    evidence that you are using the latest available version.

 3. Upgrading your dependencies regularly, whether manual or automated,
    provides an opportunity to review release notes and/or commit logs.
    This can provide meaningful context, along the way, for how that
    package is drifting over time, and to more easily recognize and
    predict how those changes might affect your app.

Whether and how often you choose to keep your app's dependencies
updated is up to you and your team's risk tolerance and the overhead
you are willing to accept in order to keep up with how your
dependencies are drifting.

I have found that, more often than not, dependency upgrades are safe.
With a healthy test suite and continuous integration, I have had very
few problems resulting from dependency upgrades.

Personally, I opt to rely on my test suite and CI pipeline to catch
any compatibility issues, and upgrade my projects' dependencies
frequently (~biweekly for updating package.json, ~monthly for
regenerating yarn.lock with updated resolutions). To make this process
easier, I wrote a tool to semi-automatically run through the whole
dependency upgrade process:

> See: https://github.com/twuni/automated-upgrades.js

Tools like Greenkeeper are nice because they will upgrade each
dependency in isolation, ensuring that any issues are caught with
fine, package-level granularity. I've found that tools like this
create a lot of noise, especially in projects with lots of
dependencies. Reviewing these PRs and triggering CI builds for every
one of these packages, dealing with merge conflicts when multiple
PRs update the lockfile, and for some of these tools to require
read/write access on the repositories they are wired to monitor...
this all, in my experience, creates more overhead and risk than
the service is designed to solve.

I've found that having a human upgrading all dependencies at once
(potentially with assistive tools like my automated-upgrades CLI)
is a good compromise. It's a relatively low overhead solution that,
more often than not, works just fine -- and when it doesn't, it's
usually not difficult to identify which dependency upgrade caused
the complication.