🔧 Define the Node.js and Yarn versions for this repo
**.tool-versions** is a file used by [asdf](https://github.com/asdf-vm/asdf)
to configure which version of which tools are required by a project.

**asdf** is a universal version manager that I strongly recommend
for all teams. If you have ever used a tool-specific version manager
like `nvm`, `n`, `rbenv`, `rvm`, or even `gradlew`, you know the
use case and the pain of working in different projects that require
different versions of tools that are typically installed system-wide.

Unlike other version managers, asdf is self-contained and unobtrusive.
to install it, clone the repo and source a script. To uninstall it,
just remove the line from your shell profile. It's everything you
ever wanted any of those other version managers to be, and it's the
only version manager you'll ever need.

Don't use asdf in your production servers or CI pipeline, though.
I mean, technically you could and it would be fine, but you would
probably be better off baking the tools you need into the base AMI
or Docker image.

Seriously, give asdf a shot. It's one of the very few tools I would
recommend to **anyone** for **any project**.