📝 Add CircleCI status badge to README
After creating a CI configuration, adding a badge to the README
communicating the CI build status serves as an indicator for the
project's health. It provides reassurance to the project's
maintainers and its consumers alike that (a) the project is using
CI automation to continuously verify and ensure its health, and
(b) *which* CI tool is being used.